requirements.txt
requests==2.32.5
beautifulsoup4==4.14.3
gspread==6.2.1
google-auth==2.48.0
google-auth-oauthlib==1.3.0
google-auth-httplib2==0.3.0
schedule==1.2.2
python-dotenv==1.2.2
